connection/requirements.txt
snowflake-connector-python
snowflake-snowpark-python==1.29.0